pyproject.toml
[tool.poetry]
authors = ["elliottjjackson <ejay01@gmail.com>"]
description = ""
name = "internet-speed-monitor"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.9"
speedtest-cli = "^2.1.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.930"
pre-commit = "^2.16.0"
pytest = "^5.2"
pytest-cov = "^3.0.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.flake8]
max-line-length = 88

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.black]
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 88
target-version = ['py38']
